git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4075 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxexpr.cpp
@@ -2581,7 +2581,7 @@ public:
 
 		FillArgs ( tMatch );
 		UdfInt_fn pFn = (UdfInt_fn) m_pCall->m_pUdf->m_fnFunc;
-		int64_t iRes = (int) pFn ( &m_pCall->m_tInit, &m_pCall->m_tArgs, &m_bError );
+		int64_t iRes = pFn ( &m_pCall->m_tInit, &m_pCall->m_tArgs, &m_bError );
 		FreeArgs();
 
 		if ( m_pProfiler )
